AB#5488 Convert project to src layout.
Python packaging and pytest generally direct people towards a `src`
layout as this has some benefits. These benefits for us are:

- easier packaging (old way mistakenly packaged `django_tests`)
- clearer separation between the `schematools` packages and its tests.
- more in line with other projects

The clearer separation between tests and the `schematools` package
immediately became obvious for the Django specific tests. They relied
on Django being configured by means of specifying
`DJANGO_SETTINGS_MODULE` that referred to a settings module in what was
the `django_tests` package. However `django_tests` wasn't longer a
Python package (it wasn't moved to `src`). Hence we needed to come up
with a different way of configuring Django for the tests.

The obvious way was to specify a `pytest_configure` function in the top
level `conftest.py` that would call out to `settings.configure`. That
wasn't sufficient as the tests used local pytest plugins (`fixtures.py`)
that imported the `schematools` packages, that in turn relied on Django
already been configured. These plugins were loaded before
`pytest_configure` could run. So we had to do away with the local
plugins. All fixtures now reside in `conftest.py` files as they should!
Using local pytest plugin just to group fixtures was a hack anyway.

We also added a `clean` target to the Makefile. And added a `dev`
options to `schematools`. This installs the `build` package that is
generally recommended to build modern Python packages.

If you have `amsterdam-schema-tools` installed using the `--editable`
feature of pip, you should do the following after having checked out
this commit:

    $ pip uninstall amsterdam-schema-tools
    $ make install

Due to the massive reordering of file location the previous editable
install might get `confused` as to what lives where.
