+++ Makefile
@@ -17,3 +17,8 @@ coverage:
 .PHONY: format
 format:
 	pre-commit run -a
+
+.PHONY: clean
+clean:
+	find . -type d -name __pycache__ -exec rm -r {} \+
+	rm -rf build dist
