+++ setup.cfg
@@ -15,6 +15,8 @@ classifiers =
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
 [options]
+package_dir =
+    = src
 packages = find:
 python_requires = >= 3.7
 install_requires =
@@ -43,8 +45,7 @@ tests_require=
     requests-mock
 
 [options.packages.find]
-exclude =
-    tests
+where = src
 
 [options.extras_require]
 tests =
@@ -71,6 +72,7 @@ django =
     django-environ
 dev =
     tbump
+    build
 
 [options.entry_points]
 console_scripts =
@@ -83,8 +85,6 @@ universal = 1
 [tool:pytest]
 testpaths =
     tests
-    tests_django
-DJANGO_SETTINGS_MODULE = tests_django.settings
 norecursedirs =
     .tox
     .git
