+++ src/schematools/contrib/django/apps.py
