+++ src/schematools/contrib/django/cli.py
