+++ src/schematools/contrib/django/management/commands/import_schemas.py
