+++ src/schematools/contrib/django/management/commands/remove_schemas.py
