+++ src/schematools/contrib/django/managers.py
