+++ src/schematools/datasetcollection.py
