+++ src/schematools/exceptions.py
