+++ src/schematools/importer/geojson.py
