+++ src/schematools/importer/ndjson.py
