+++ src/schematools/introspect/db.py
