+++ src/schematools/introspect/geojson.py
