+++ src/schematools/introspect/utils.py
