+++ src/schematools/maps/generators/mapfile.py
