+++ src/schematools/maps/interfaces/mapfile/types.py
