+++ src/schematools/utils.py
