+++ src/schematools/validation.py
