--- tests_django/conftest.py
@@ -1,4 +1,0 @@
-pytest_plugins = [
-    "tests.fixtures",
-    "tests_django.fixtures",
-]
